Documentation and source code of the ICEMap-R software
Documentation (/doc) and source code (/py) relevant to the River Ice Jam
mapping Tool (ICEMap-R) that are hosted in this GIT repository are
covered by a mozilla license.

Copyright (c)2016, Institut National de la Recherche Scientifique (INRS)

Licensed under the Mozilla Public License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

https://www.mozilla.org/en-US/MPL/2.0/

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.